wave4(phase1): extract Django-free run-lifecycle core into installable canopy_runs package
Prereq for the ACE migration (Option 2 — ace-web imports the framework code):
- New packages/canopy_runs/ — a pip-installable, Django-FREE library: the read
  model (schemas), the RunStore Protocol + InMemoryRunStore, and the full Drive
  adapter (DriveClient/parsers/DriveRunStore/GoogleDriveClient, registry injected
  not read from settings). google SDK under an optional [drive] extra, imported
  lazily.
- apps/agent_runs is now a thin Django wrapper: models + DbRunStore + resolver +
  REST, importing canopy_runs (schemas re-exported for back-compat).
- Library imports with no Django in sys.modules; uv path-dependency wires it.
- 505 (Django) + 129 (lib) = 634 tests, matching baseline; boundary clean.
