+++ pyproject.toml
@@ -22,6 +22,11 @@ dependencies = [
     "google-api-python-client>=2.120,<3.0",
     "google-auth>=2.29,<3.0",
     "fastmcp>=3.3,<4.0",
+    # The Django-free run-lifecycle core (read model + RunStore Protocol + Drive
+    # adapter). Lives in-repo at packages/canopy_runs; resolved as an editable
+    # path dep via [tool.uv.sources] below. Django's DbRunStore (apps/agent_runs)
+    # implements the same Protocol over the ORM.
+    "canopy-runs",
 ]
 
 [project.optional-dependencies]
@@ -36,6 +41,13 @@ dev = [
 [build-system]
 requires = ["setuptools>=68.0", "wheel"]
 build-backend = "setuptools.build_meta"
+
+[tool.uv.sources]
+# The run-lifecycle core ships in-repo as a standalone installable package; wire
+# it as an editable path dep so `from canopy_runs import ...` resolves under
+# `uv run` without publishing to an index. A separate project (ace-web) installs
+# the same package from its own source/index.
+canopy-runs = { path = "packages/canopy_runs", editable = true }
 
 [tool.setuptools.packages.find]
 include = ["apps*", "config*"]
@@ -52,3 +64,7 @@ select = ["E", "F", "I", "N", "W", "UP"]
 DJANGO_SETTINGS_MODULE = "config.settings.test"
 python_files = ["tests.py", "test_*.py", "*_tests.py"]
 asyncio_mode = "auto"
+# The standalone packages/canopy_runs has its OWN plain-pytest suite (no Django,
+# no DB); it is run from its own dir, not as part of the Django suite. Ignore it
+# here so the Django runner doesn't try to collect it under DJANGO_SETTINGS.
+addopts = "--ignore=packages"
